requirements.txt
numpy
pandas
scipy
scikit-learn
xgboost
matplotlib
seaborn
shap
umap-learn
combatlearn
tqdm
maldiamrkit